==> notebooks/0_3_predict_transactions.ipynb <==
library(zoo)
library(data.table)
library(lubridate)
library(forecast)

# install.packages('forecast')

print(version)

print(R.home())

train_data <- as.data.table(read.csv("../cache/transactions/tr_25_20160101.csv"))


test_data <- as.data.table(read.csv("../cache/transactions/te_25_20160101.csv"))

train_data$date <- ymd(train_data$date)

head(train_data)

tail(train_data)

train_data$date <- as.Date(train_data$date)

date <- as.Date(seq(as.Date("2013-01-01"), as.Date("2015-12-31"), by="days"))

new_train <- as.data.frame(date)

head(new_train)

new_train <- merge(new_train,train_data,by="date",all.x=T)

table(is.na(new_train))

new_train[is.na(new_train)] <- 0

head(new_train)

x.ts = ts(new_train$transactions, freq=365, start=c(2013))

hw1 <- HoltWinters(x.ts,seasonal = "additive",beta = FALSE)

head(test_data)

nrow(test_data)

nrow(new_train)

test_data$transactions <- predict(hw1,n.ahead = 1)

table(is.na(test_data))

nrow(test_data)

head(test_data)

test_data$transactions = as.integer(test_data$transactions)

pred = test_data$transactions

pred

head(test_data)

trans = data.frame()

trans1 = data.frame('2016-01-03', 52, 0)

trans = rbind(trans, trans1)

names(trans) = c('date', 'store_nbr', 'transactions')

trans

trans = data.frame()

write.csv(test_data,file ="../cache/transactions_25_20160101.csv",row.names = FALSE)

paste('../cache/transactions/tr_',toString(10),'_20160103.csv', sep='')

for (i in 1:54) {
    if (i == 52) {
        next
    }
    fn_tr = paste('../cache/transactions/tr_',toString(i),'_20160103.csv', sep='')
    fn_te = paste('../cache/transactions/te_',toString(i),'_20160103.csv', sep='')
    
    train_data <- as.data.table(read.csv(fn_tr))
    test_data <- as.data.table(read.csv(fn_te))
    train_data$date <- ymd(train_data$date)
    train_data$date <- as.Date(train_data$date)
    date <- as.Date(seq(as.Date("2013-01-01"), as.Date("2016-01-02"), by="days"))
    new_train <- as.data.frame(date)
    new_train <- merge(new_train,train_data,by="date",all.x=T)
    new_train[is.na(new_train)] <- 0
    x.ts = ts(new_train$transactions, freq=365, start=c(2013))
    hw1 <- HoltWinters(x.ts,seasonal = "additive",beta = FALSE)
    test_data$transactions <- predict(hw1,n.ahead = 1)
    test_data$transactions = as.integer(test_data$transactions)
    pred  = test_data$transactions
    new_test = data.frame('2016-01-03', i, pred)
    trans = rbind(trans, new_test)
}

head(trans)

names(trans) = c('date', 'store_nbr', 'transactions')

tail(trans)

head(trans)

table(is.na(trans))

write.csv(trans,file ="../cache/transactions_20160103.csv",row.names = FALSE)

# now other missing data

l = c('2013-06-19_10',
 '2013-06-19_35',
 '2013-06-19_43',
 '2013-06-19_54',
 '2014-01-02_32',
 '2014-03-24_25',
 '2016-01-02_1',
 '2016-01-02_10',
 '2016-01-02_11',
 '2016-01-02_12',
 '2016-01-02_13',
 '2016-01-02_14',
 '2016-01-02_15',
 '2016-01-02_17',
 '2016-01-02_18',
 '2016-01-02_19',
 '2016-01-02_3',
 '2016-01-02_4',
 '2016-01-02_5',
 '2016-01-02_6',
 '2016-01-02_7',
 '2016-01-02_8',
 '2016-01-02_9',
 '2016-01-04_10',
 '2016-01-04_11',
 '2016-01-04_12',
 '2016-01-04_13',
 '2016-01-04_14',
 '2016-01-04_15',
 '2016-01-04_16',
 '2016-01-04_17',
 '2016-01-04_18',
 '2016-01-04_19',
 '2016-01-04_2',
 '2016-01-04_20',
 '2016-01-04_21',
 '2016-01-04_22',
 '2016-01-04_27',
 '2016-01-04_28',
 '2016-01-04_29',
 '2016-01-04_3',
 '2016-01-04_35',
 '2016-01-04_36',
 '2016-01-04_4',
 '2016-01-04_40',
 '2016-01-04_41',
 '2016-01-04_42',
 '2016-01-04_43',
 '2016-01-04_44',
 '2016-01-04_45',
 '2016-01-04_46',
 '2016-01-04_47',
 '2016-01-04_48',
 '2016-01-04_49',
 '2016-01-04_50',
 '2016-01-04_51',
 '2016-01-04_53',
 '2016-01-04_54',
 '2016-01-04_6',
 '2016-01-04_7',
 '2016-01-04_8',
 '2016-01-04_9',
 '2016-09-27_23',
 '2016-09-27_7')

length(l)

date_and_store= strsplit('2016-09-27_7', '_')[[1]]

date_and_store[2]

date <- date_and_store[1]
date <- as.Date(date) -1
print(date)

date <- date_and_store[1]
date <- gsub('-','',date)
i <- date_and_store[2]
paste('../cache/transactions/tr_',toString(i),'_', toString(date), '.csv', sep='')

trans = data.frame()

for (el in l) {
    print(el)
    date_and_store = strsplit(el, '_')[[1]]
    date1 <- date_and_store[1]
    date2 <- gsub('-','',date1)
    i <- date_and_store[2]
    fn_tr = paste('../cache/transactions/tr_',toString(i),'_', toString(date2), '.csv', sep='')
    fn_te = paste('../cache/transactions/te_',toString(i),'_', toString(date2), '.csv', sep='')
    print(fn_tr)
    train_data <- as.data.table(read.csv(fn_tr))
    print(nrow(train_data))
    test_data <- as.data.table(read.csv(fn_te))
    train_data$date <- ymd(train_data$date)
    train_data$date <- as.Date(train_data$date)
    date <- as.Date(seq(as.Date("2013-01-01"), as.Date(date1)-1, by="days"))
    new_train <- as.data.frame(date)
    new_train <- merge(new_train,train_data,by="date",all.x=T)
    new_train[is.na(new_train)] <- 0
    x.ts = ts(new_train$transactions, freq=30, start=c(2013))
    hw1 <- HoltWinters(x.ts,seasonal = "additive",beta = FALSE)
    test_data$transactions <- predict(hw1,n.ahead = 1)
    test_data$transactions = as.integer(test_data$transactions)
    pred  = test_data$transactions
    new_test = data.frame(date1, i, pred)
    trans = rbind(trans, new_test)
}

nrow(trans)

names(trans) = c('date', 'store_nbr', 'transactions')

write.csv(trans,file ="../cache/transactions_missing.csv",row.names = FALSE)

# now for test data

trans = data.frame()

for (i in 1:54) {
    
    fn_tr = paste('../cache/transactions/tr_all_',toString(i),'.csv', sep='')
    fn_te = paste('../cache/transactions/te_all_',toString(i),'.csv', sep='')
    
    train_data <- as.data.table(read.csv(fn_tr))
    test_data <- as.data.table(read.csv(fn_te))
    train_data$date <- ymd(train_data$date)
    train_data$date <- as.Date(train_data$date)
    date <- as.Date(seq(as.Date("2013-01-01"), as.Date("2017-08-15"), by="days"))
    new_train <- as.data.frame(date)
    new_train <- merge(new_train,train_data,by="date",all.x=T)
    new_train[is.na(new_train)] <- 0
    x.ts = ts(new_train$transactions, freq=365, start=c(2013))
    hw1 <- HoltWinters(x.ts,seasonal = "additive",beta = FALSE)
    test_data$transactions <- predict(hw1,n.ahead = 16)
    test_data$transactions = as.integer(test_data$transactions)
    
    test_data$store_nbr = toString(i)
    trans = rbind(trans, as.data.frame(test_data))
}

names(trans) = c('date', 'transactions', 'store_nbr')

nrow(trans)

head(trans)

16 * 54

write.csv(trans,file ="../cache/test_transactions.csv",row.names = FALSE)
