==> notebooks/1_0_prepare_splits and_counts.ipynb <==
import argparse, csv, sys, collections
import os
from tqdm import tqdm
import feather
import pandas as pd
from joblib import Parallel, delayed
import numpy as np

# from https://github.com/alexeygrigorev/outbrain-click-prediction-kaggle/blob/master/1_svm_data.py
def append_to_csv(batch, csv_file):
    props = dict(encoding='utf-8', index=False)
    if not os.path.exists(csv_file):
        batch.to_csv(csv_file, **props)
    else:
        batch.to_csv(csv_file, mode='a', header=False, **props)

def delete_file_if_exists(filename):
    if os.path.exists(filename):
        os.remove(filename)

def chunk_dataframe(df, n):
    for i in range(0, len(df), n):
        yield df.iloc[i:i+n]


def prepare_batch(batch):
    batch = batch.reset_index(drop=1)
    
    return batch

df_train  = feather.read_dataframe('../cache/train10_t.feather')

df_train.head()

df_train.shape

df_train.columns

df_train = df_train.rename(columns = {'family':'item_family', 'class': 'item_class', 'type': 'h_type', 
                                      'description': 'h_desc'})

df_train.head()

df_train.columns

df_train['store_item_nbr'] = df_train['store_nbr'].astype('str') + '_' + df_train['item_nbr'].astype('str')

delete_file_if_exists('../cache/train10_t.csv')

for batch in tqdm(chunk_dataframe(df_train, n=1000000)):
    batch = prepare_batch(batch)
    append_to_csv(batch, '../cache/train10_t.csv')


f_cats = ['store_item_nbr', 'onpromotion', 'item_family', 'item_class', 'city', 'state', 'store_type', 
          'cluster', 'dom', 'dow', 'h_type', 'pd']

counts = collections.defaultdict(lambda : [0, 0, 0])



def count_values_in_row(csv_path):
    for i, row in tqdm(enumerate(csv.DictReader(open(csv_path)), start=2)):
        label = float(row['unit_sales'])
        for j in f_cats:
            field = j
            value = row[field]
            if label <= 0.:
                counts[field+','+value][0] += 1
            else:
                counts[field+','+value][1] += 1
            counts[field+','+value][2] += 1
        if i % 1000000 == 0:
            sys.stderr.write('{0}m\n'.format(int(i/1000000)))

def count_test_values_in_row(csv_path):
    for i, row in tqdm(enumerate(csv.DictReader(open(csv_path)), start=2)):
#         label = float(row['unit_sales'])
        for j in f_cats:
            field = j
            value = row[field]
            counts[field+','+value][0] += 1
            counts[field+','+value][1] += 1
            counts[field+','+value][2] += 1
        if i % 1000000 == 0:
            sys.stderr.write('{0}m\n'.format(int(i/1000000)))
            
def print_results(dest_csv):
    fp = open(dest_csv, 'w')
    fp.write('Field,Value,Neg,Pos,Total,Ratio\n')
    for key, (neg, pos, total) in sorted(counts.items(), key=lambda x: x[1][2]):
        if total < 10:
            continue
        ratio = round(float(pos)/total, 5)
        counts_str = key+','+str(neg)+','+str(pos)+','+str(total)+','+str(ratio) + '\n'
        fp.write(counts_str)
    fp.close()

count_values_in_row('../cache/train10_t.csv')

print_results('../cache/counts_train10_t.csv')

df_test  = feather.read_dataframe('../cache/test7_t.feather')

df_test = df_test.rename(columns = {'family':'item_family', 'class': 'item_class', 'type': 'h_type', 
                                      'description': 'h_desc'})

df_test['store_item_nbr'] = df_test['store_nbr'].astype('str') + '_' + df_test['item_nbr'].astype('str')

delete_file_if_exists('../cache/test7_t.csv')

for batch in tqdm(chunk_dataframe(df_test, n=1000000)):
    batch = prepare_batch(batch)
    append_to_csv(batch, '../cache/test7_t.csv')


counts = collections.defaultdict(lambda : [0, 0, 0])

count_test_values_in_row('../cache/test7_t.csv')

print_results('../cache/counts_test7_t.csv')

df = pd.read_csv('../cache/counts_train10_t.csv')

df.head()

df.shape

df1 = df[df.Total >= 12000000]

df1.shape


df1

df1.to_csv('../cache/counts_dense_train10_t.csv')

df_train['dense_count'] = 0

df_train = df_train.drop('dense_count', axis=1)

df_train = df_train.set_index('date')

df_train.head()

df_train.shape

df_train = df_train.reset_index()

df_train.shape

df_train.head()

past_date = np.datetime64('2015-01-01')

df_train = df_train[(df_train.date.apply(np.datetime64) >= past_date)]

df_train.shape

df_train.head()

df_train.tail()

import os
x = {i for i in range(20)}
os.sched_setaffinity(0, x)

# now split in to various stores

len(set(df_train.store_nbr))

set(range(1,55)) - set(df_train.store_nbr)

set(df_train.store_nbr) - set(range(1,55))

import gc
gc.enable()

for i in tqdm(range(1,55)):
    df = df_train[df_train.store_nbr == i]
    fn = '../cache/train10_t_store{}.csv'.format(i)
    df.to_csv(fn, index=False)
    del df

gc.collect()

set(range(1,55)) - set(df_test.store_nbr)

len(set(df_test.store_nbr))

df_test.shape

df_train.shape

for i in tqdm(range(1,55)):
    df = df_test[df_test.store_nbr == i]
    fn = '../cache/test7_t_store{}.csv'.format(i)
    df.to_csv(fn, index=False)
    del df

gc.collect()

df_test  = pd.read_csv('../cache/test7_t.csv')

df_test.T

df_sub = pd.read_csv('../submissions/lgb.csv')

df_sub.head()

df_test = pd.merge(df_test, df_sub, how='left', on='id')

df_test.isnull().sum()

df_test.shape

for i in tqdm(range(1,55)):
    df = df_test[df_test.store_nbr == i]
    fn = '../cache/bkup2/test7_1_t_store{}.csv'.format(i)
    df.to_csv(fn, index=False, float_format='%.5f')
    del df
